+++ src/main/java/com/community/dogcat/controller/board/scrap/ScrapController.java
@@ -1,5 +1,7 @@
 package com.community.dogcat.controller.board.scrap;
 
+import static org.springframework.http.HttpStatus.*;
+
 import java.util.HashMap;
 import java.util.Map;
 
@@ -17,6 +19,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.community.dogcat.controller.BaseController;
+import com.community.dogcat.domain.PostLike;
+import com.community.dogcat.domain.Scrap;
 import com.community.dogcat.dto.board.scrap.ScrapDTO;
 import com.community.dogcat.jwt.JWTUtil;
 import com.community.dogcat.service.board.scrap.ScrapService;
@@ -45,7 +49,11 @@ public class ScrapController extends BaseController {
 		// 모델에서 사용자 정보를 가져옴
 		String userId = (String)model.getAttribute("username");
 
-		scrapDTO.setUserId(userId);
+		// 로그인 사용자 확인
+		if (userId == null) {
+			log.error("ScrapController register Error : 401 Unauthorized");
+			return ResponseEntity.status(UNAUTHORIZED).build(); // 로그인되지 않은 경우 401 오류
+		}
 
 		Map<String, Long> response = new HashMap<>();
 
@@ -62,6 +70,18 @@ public class ScrapController extends BaseController {
 
 		// 모델에서 사용자 정보를 가져옴
 		String userId = (String)model.getAttribute("username");
+
+		// 로그인 사용자 확인
+		if (userId == null) {
+			log.error("ScrapController Delete Error : 401 Unauthorized");
+			return ResponseEntity.status(UNAUTHORIZED).build(); // 로그인되지 않은 경우 401 오류
+		}
+
+		// 보관 정보 조회
+		if (scrapNo == null) {
+			log.error("ScrapController Delete Error : 404 Not Found");
+			return ResponseEntity.status(NOT_FOUND).build(); // 좋아요가 존재하지 않는 경우 404 오류
+		}
 
 		scrapService.delete(scrapNo, userId);
